mcap examples: add live pose viz + refactor CloudXRLauncher + consolidate viz classes
- Add live_hand.py, live_controller.py, live_full_body.py: real-time viser
  visualization of XR pose data using CloudXRLauncher + TeleopSession (no MCAP)
- Refactor record_* scripts to use CloudXRLauncher so the CloudXR runtime
  starts automatically; add --accept-eula / --install-dir args
- Move HandViz, ControllerViz, FullBodyViz and shared color constants into
  common.py; replay_* scripts now import from there instead of duplicating
