+++ examples/mcap_record_replay/python/replay_controller.py
@@ -23,19 +23,23 @@ import sys
 import time
 from pathlib import Path
 
-import numpy as np
 import viser
 from mcap.reader import make_reader
 
 from isaacteleop.deviceio import McapReplayConfig
-from isaacteleop.retargeting_engine.tensor_types.indices import ControllerInputIndex
 from isaacteleop.teleop_session_manager import (
     SessionMode,
     TeleopSession,
     TeleopSessionConfig,
 )
 
-from common import build_controller_pipeline
+from common import (
+    ControllerViz,
+    LEFT_COLOR,
+    RIGHT_COLOR,
+    build_controller_pipeline,
+    controller_state,
+)
 
 
 def mcap_duration_s(path: Path) -> float:
@@ -56,11 +60,6 @@ def mcap_duration_s(path: Path) -> float:
         return (stats.message_end_time - stats.message_start_time) / 1e9
 
 
-LEFT_COLOR = (0.25, 0.85, 0.35)
-RIGHT_COLOR = (0.35, 0.55, 0.95)
-INVALID_COLOR = (1.0, 0.0, 0.0)
-
-
 def resolve_mcap(path_arg: str | None) -> Path:
     if path_arg:
         path = Path(path_arg)
@@ -76,175 +75,6 @@ def resolve_mcap(path_arg: str | None) -> Path:
             "Run record_controller.py first or pass a path."
         )
     return max(candidates, key=lambda p: p.stat().st_mtime)
-
-
-def _segment(start: np.ndarray, end: np.ndarray) -> np.ndarray:
-    return np.stack([start, end], axis=0).astype(np.float32)
-
-
-class ControllerViz:
-    """Per-controller viser handles (3D pose + live input-state HUD)."""
-
-    def __init__(
-        self,
-        server: viser.ViserServer,
-        name: str,
-        color: tuple[float, float, float],
-    ):
-        self.color = np.array(color, dtype=np.float32)
-        zero_pt = np.zeros((1, 3), dtype=np.float32)
-        zero_seg = np.zeros((0, 2, 3), dtype=np.float32)
-        zero_seg_colors = np.zeros((0, 2, 3), dtype=np.float32)
-
-        self.aim = server.scene.add_point_cloud(
-            name=f"/{name}/aim",
-            points=zero_pt,
-            colors=np.tile(self.color, (1, 1)),
-            point_size=0.015,
-        )
-        self.grip = server.scene.add_point_cloud(
-            name=f"/{name}/grip",
-            points=zero_pt,
-            colors=np.tile(self.color, (1, 1)),
-            point_size=0.015,
-        )
-        self.ray = server.scene.add_line_segments(
-            name=f"/{name}/ray",
-            points=zero_seg,
-            colors=zero_seg_colors,
-            line_width=2.0,
-        )
-
-        with server.gui.add_folder(name):
-            self.hud_tracking = server.gui.add_checkbox("tracked", False, disabled=True)
-            self.hud_aim_valid = server.gui.add_checkbox(
-                "aim_valid", False, disabled=True
-            )
-            self.hud_grip_valid = server.gui.add_checkbox(
-                "grip_valid", False, disabled=True
-            )
-            self.hud_stick = server.gui.add_vector2(
-                "thumbstick_xy",
-                initial_value=(0.0, 0.0),
-                min=(-1.0, -1.0),
-                max=(1.0, 1.0),
-                disabled=True,
-            )
-            self.hud_trigger_value = server.gui.add_number(
-                "trigger",
-                initial_value=0.0,
-                min=0.0,
-                max=1.0,
-                step=0.01,
-                disabled=True,
-            )
-            self.hud_trigger = server.gui.add_progress_bar(0.0)
-            self.hud_squeeze_value = server.gui.add_number(
-                "squeeze",
-                initial_value=0.0,
-                min=0.0,
-                max=1.0,
-                step=0.01,
-                disabled=True,
-            )
-            self.hud_squeeze = server.gui.add_progress_bar(0.0)
-            self.hud_primary = server.gui.add_checkbox(
-                "primary_click", False, disabled=True
-            )
-            self.hud_secondary = server.gui.add_checkbox(
-                "secondary_click", False, disabled=True
-            )
-            self.hud_stick_click = server.gui.add_checkbox(
-                "thumbstick_click", False, disabled=True
-            )
-            self.hud_menu_click = server.gui.add_checkbox(
-                "menu_click", False, disabled=True
-            )
-
-    def update(self, state: dict) -> None:
-        aim_valid: bool = state["aim_valid"]
-        grip_valid: bool = state["grip_valid"]
-        aim_pos: np.ndarray | None = state["aim_pos"]
-        grip_pos: np.ndarray | None = state["grip_pos"]
-
-        self.hud_tracking.value = state["tracked"]
-        self.hud_aim_valid.value = aim_valid
-        self.hud_grip_valid.value = grip_valid
-        self.hud_stick.value = state["thumbstick_xy"]
-        self.hud_trigger.value = max(0.0, min(1.0, state["trigger"]))
-        self.hud_trigger_value.value = state["trigger"]
-        self.hud_squeeze.value = max(0.0, min(1.0, state["squeeze"]))
-        self.hud_squeeze_value.value = state["squeeze"]
-        self.hud_primary.value = state["primary_click"]
-        self.hud_secondary.value = state["secondary_click"]
-        self.hud_stick_click.value = state["thumbstick_click"]
-        self.hud_menu_click.value = state["menu_click"]
-
-        if aim_valid and aim_pos is not None:
-            self.aim.points = aim_pos.reshape(1, 3).astype(np.float32)
-            self.aim.colors = np.tile(self.color, (1, 1))
-        else:
-            self.aim.points = np.zeros((1, 3), dtype=np.float32)
-            self.aim.colors = np.tile(INVALID_COLOR, (1, 1))
-
-        if grip_valid and grip_pos is not None:
-            self.grip.points = grip_pos.reshape(1, 3).astype(np.float32)
-            self.grip.colors = np.tile(self.color, (1, 1))
-        else:
-            self.grip.points = np.zeros((1, 3), dtype=np.float32)
-            self.grip.colors = np.tile(INVALID_COLOR, (1, 1))
-
-        if aim_valid and grip_valid and aim_pos is not None and grip_pos is not None:
-            seg = _segment(grip_pos, aim_pos).reshape(1, 2, 3)
-            self.ray.points = seg
-            self.ray.colors = np.tile(self.color, (1, 2, 1))
-        else:
-            self.ray.points = np.zeros((0, 2, 3), dtype=np.float32)
-            self.ray.colors = np.zeros((0, 2, 3), dtype=np.float32)
-
-
-def _controller_state(controller):
-    if controller.is_none:
-        return {
-            "aim_pos": None,
-            "grip_pos": None,
-            "aim_valid": False,
-            "grip_valid": False,
-            "trigger": 0.0,
-            "squeeze": 0.0,
-            "thumbstick_xy": (0.0, 0.0),
-            "primary_click": False,
-            "secondary_click": False,
-            "thumbstick_click": False,
-            "menu_click": False,
-            "tracked": False,
-        }
-
-    aim_valid = bool(controller[ControllerInputIndex.AIM_IS_VALID])
-    grip_valid = bool(controller[ControllerInputIndex.GRIP_IS_VALID])
-    return {
-        "aim_pos": np.asarray(
-            controller[ControllerInputIndex.AIM_POSITION], dtype=np.float32
-        ),
-        "grip_pos": np.asarray(
-            controller[ControllerInputIndex.GRIP_POSITION], dtype=np.float32
-        ),
-        "aim_valid": aim_valid,
-        "grip_valid": grip_valid,
-        "trigger": float(controller[ControllerInputIndex.TRIGGER_VALUE]),
-        "squeeze": float(controller[ControllerInputIndex.SQUEEZE_VALUE]),
-        "thumbstick_xy": (
-            float(controller[ControllerInputIndex.THUMBSTICK_X]),
-            float(controller[ControllerInputIndex.THUMBSTICK_Y]),
-        ),
-        "primary_click": float(controller[ControllerInputIndex.PRIMARY_CLICK]) > 0.5,
-        "secondary_click": float(controller[ControllerInputIndex.SECONDARY_CLICK])
-        > 0.5,
-        "thumbstick_click": float(controller[ControllerInputIndex.THUMBSTICK_CLICK])
-        > 0.5,
-        "menu_click": float(controller[ControllerInputIndex.MENU_CLICK]) > 0.5,
-        "tracked": aim_valid or grip_valid,
-    }
 
 
 def run_once(
@@ -266,11 +96,9 @@ def run_once(
         start = time.time()
         while time.time() - start < duration_s:
             result = session.step()
-            left = result["controller_left"]
-            right = result["controller_right"]
 
-            l_state = _controller_state(left)
-            r_state = _controller_state(right)
+            l_state = controller_state(result["controller_left"])
+            r_state = controller_state(result["controller_right"])
 
             viz_left.update(l_state)
             viz_right.update(r_state)
